Move dummy astCreator
