+++ src/stepper/stepper.ts
@@ -22,7 +22,7 @@ import {
   dummyProgram,
   dummyStatement,
   dummyVariableDeclarator
-} from '../utils/dummyAstCreator'
+} from '../utils/ast/dummyAstCreator'
 import { evaluateBinaryExpression, evaluateUnaryExpression } from '../utils/operators'
 import * as rttc from '../utils/rttc'
 import {
